Initialize a Categorical likelihood node

This function constructs a likelihood object to represent a categorical likelihood node.

Usage

Categorical(K = NULL, input_dim = NULL)

Arguments

K

an integer specifying the number of classes in the training data. If K is set to NULL (the default), its value will be inferred from the training data when running dgp(). Defaults to NULL.

input_dim

a vector that specifies which GP node outputs feed into this likelihood node. It should be either:

A vector of length 1 (for binary output), or

A vector of length K (for K-class output), containing the indices of one or K GP nodes in the feeding layer.

If input_dim is set to NULL (the default), all GP node outputs from the feeding layer feed into the likelihood node. In this case, you must ensure that the feeding layer contains exactly one or K GP nodes. Defaults to NULL.

Value

A 'python' object to represent a categorical likelihood node.

Details

See further examples and tutorials at https://mingdeyu.github.io/dgpsi-R/.

Note

The categorical likelihood node can only be linked to one or K feeding GP nodes.

Examples

if (FALSE) { # \dontrun{

# Check https://mingdeyu.github.io/dgpsi-R/ for examples
# on how to customize DGP structures using Categorical().
} # }