dgpsi: Interface to 'dgpsi' for Deep and Linked Gaussian Process Emulations

Interface to the 'python' package 'dgpsi' for Gaussian process, deep Gaussian process, and linked deep Gaussian process emulations of computer models and networks using stochastic imputation (SI). The implementations follow Ming & Guillas (2021) doi:10.1137/20M1323771 and Ming, Williamson, & Guillas (2023) doi:10.1080/00401706.2022.2124311 and Ming & Williamson (2023) arXiv:2306.01212. To get started with the package, see https://mingdeyu.github.io/dgpsi-R/.

See also

Author

Maintainer: Deyu Ming deyu.ming.16@ucl.ac.uk [copyright holder]

Authors:

Daniel Williamson d.williamson@exeter.ac.uk