Validation plots of a sequential design

This function draws validation plots of the sequential design of a (D)GP emulator or a bundle of (D)GP emulators.

Usage

draw(object, ...)

# S3 method for class 'gp'
draw(object, type = "rmse", log = FALSE, ...)

# S3 method for class 'dgp'
draw(object, type = "rmse", log = FALSE, ...)

# S3 method for class 'bundle'
draw(object, emulator = 1, type = "rmse", log = FALSE, ...)

Arguments

object

can be one of the following emulator classes:

the S3 class gp.

the S3 class dgp.

the S3 class bundle.

...

N/A.

type

either "rmse", for the trace plot of RMSEs or customized evaluating metrics of emulators constructed during the sequential designs, or "design", for visualizations of input designs created by the sequential design procedure. Defaults to "rmse".

log

a bool that indicates whether to plot RMSEs or customized evaluating metrics in log-scale if type = "rmse". Defaults to FALSE.

emulator

the index of the emulator packed in object if object is an instance of the bundle class.

Value

A patchwork object.

Details

See further examples and tutorials at https://mingdeyu.github.io/dgpsi-R/.

Note

If a customized evaluating function is provided to design() and the function returns a single evaluating metric value when object is an instance of the bundle class, the value of emulator has no effects on the plot when type = "rmse".

Examples

if (FALSE) { # \dontrun{

# See design() for an example.
} # }