Gaussian process emulator construction

This function builds and trains a GP emulator.

Usage

gp(
  X,
  Y,
  struc = NULL,
  name = "sexp",
  lengthscale = rep(0.2, ncol(X)),
  nugget_est = FALSE,
  nugget = 1e-06,
  verb = 1,
  internal_input_idx = NULL,
  linked_idx = NULL
)

Arguments

X

a matrix where each row is an input data point and each column is an input dimension.

Y

a matrix with only one column and each row being an input data point.

struc

an object produced by kernel() that gives a user-defined GP specifications. When struc = NULL, the GP specifications are automatically generated using information provided in name, lengthscale, nugget_est, nugget, and internal_input_idx. Defaults to NULL.

name

kernel function to be used. Either "sexp" for squared exponential kernel or "matern2.5" for Matérn-2.5 kernel. Defaults to "sexp". This argument is only used when struc = NULL.

lengthscale

initial values of lengthscales in the kernel function. It can be a single numeric value or a vector:

if it is a single numeric value, it is assumed that kernel functions across input dimensions share the same lengthscale;

if it is a vector (which must have a length of ncol(X)), it is assumed that kernel functions across input dimensions have different lengthscales.

Defaults to a vector of 0.2. This argument is only used when struc = NULL.

nugget_est

a bool indicating if the nugget term is to be estimated:

FALSE: the nugget term is fixed to $1e-6$ for deterministic emulation where the emulator interpolates the training data points.

TRUE: the nugget term will be estimated for stochastic emulation where computer model outputs are assumed to follow a homogeneous Gaussian distribution.

Defaults to FALSE. This argument is only used when struc = NULL.

nugget

the initial nugget value. If nugget_est = FALSE, the assigned value is fixed during the training. Defaults to 1e-6.

verb

an integer indicating the level of information to be printed during the function execution:

2: trace information on GP initialization and a summary table of the initialized GP emulator.

1: all information as in 2 except for the summary table.

0: no trace information and the summary table.

When verb = 2, you will have a chance to check the specified GP emulator (especially when a customized struc is provided) and decide if you want to proceed to training. Defaults to 1.

internal_input_idx

the column indices of X that are generated by the linked emulators in the feeding layer. Set internal_input_idx = NULL if the GP emulator is not linked to any emulator (e.g., the GP emulator is in the first layer of a system) or all columns in X are generated by the linked emulators in the feeding layer. Defaults to NULL. This argument is only used when struc = NULL.

linked_idx

the indices of columns in the pooled output matrix (formed by column-combined outputs of all emulators in the feeding layer) that will feed into the GP emulator. The length of linked_idx shall equal to the length of internal_input_idx when internal_input_idx is not NULL. Set linked_idx = NULL if the GP emulator is not intended for linked emulations. If the GP emulator is in the first layer of a system, linked_idx gives the column indices of the global input (formed by column-combining all input matrices of emulators in the first layer) that the GP emulator will use. Defaults to NULL.

Value

An S3 class named gp that can be used by

predict() for GP predictions.

lgp() to construct linked (D)GP emulators.

Details

See examples in Articles at https://mingdeyu.github.io/dgpsi-R/.