Locate the next design point for a (D)GP emulator or a bundle of (D)GP emulators

This function searches from a candidate set to locate the next design point to be added to a (D)GP or a bundle of (D)GP emulators.

Usage

locate(
  object,
  x_cand,
  n_cand,
  limits,
  batch_size,
  method,
  nugget_s,
  verb,
  cores,
  ...
)

# S3 method for gp
locate(
  object,
  x_cand = NULL,
  n_cand = 200,
  limits = NULL,
  batch_size = 1,
  method = "ALM",
  nugget_s = 1,
  verb = TRUE,
  cores = 1,
  ...
)

# S3 method for dgp
locate(
  object,
  x_cand = NULL,
  n_cand = 200,
  limits = NULL,
  batch_size = 1,
  method = "ALM",
  nugget_s = 1,
  verb = TRUE,
  cores = 1,
  threading = FALSE,
  aggregate = NULL,
  ...
)

# S3 method for bundle
locate(
  object,
  x_cand = NULL,
  n_cand = 200,
  limits = NULL,
  batch_size = 1,
  method = "ALM",
  nugget_s = 1,
  verb = TRUE,
  cores = 1,
  threading = FALSE,
  aggregate = NULL,
  ...
)

Arguments

object

can be one of the following:

the S3 class gp.

the S3 class dgp.

the S3 class bundle.

x_cand

a matrix (with each row being a design point and column being an input dimension) that gives a candidate set in which the next design point is determined. If x_cand = NULL, the candidate set will be generated using n_cand and limits. Defaults to NULL.

n_cand

an integer that gives the size of the candidate set in which the next design point is determined. This argument is used when x_cand = NULL. Defaults to 200

limits

a two-column matrix that gives the ranges of each input dimension, or a vector of length two if there is only one input dimension. If a vector is provided, it will be converted to a two-column row matrix. The rows of the matrix correspond to input dimensions, and its first and second columns correspond to the minimum and maximum values of the input dimensions. If limits = NULL, the ranges of input dimensions will be determined from the training data contained in object. This argument is used when x_cand = NULL. Defaults to NULL.

batch_size

an integer that gives the number of design points (for each simulator output dimension if aggregate = NULL) to be chosen. Defaults to 1.

method

the criterion used to locate the next design point. There are two built-in criterion: ALM ("ALM") or MICE ("MICE"), see references below. A customized criterion can also be used by supplying an R function to this argument. If a function is provided, it must be specified in the following basic form:

the first argument is an emulator object that can be either an instance of the gp class (produced by gp()) or dgp class (produced by dgp()).

the second argument is a matrix that represents a set of design points at which the scores of the criterion are calculated. Each row of the matrix gives one design point.

the output is a matrix with rows corresponding to different design points and columns corresponding to different output dimensions.

Defaults to "ALM".

nugget_s

the value of the smoothing nugget term used when method = "MICE". Defaults to 1.0.

verb

a bool indicating if the trace information will be printed during the function execution. Defaults to TRUE.

cores

the number of cores/workers to be used for the criterion calculation. If set to NULL, the number of cores is set to (max physical cores available - 1). Defaults to 1.

...

any arguments (with names different from those of arguments used in locate()) that are used by aggregate and method (if a function is supplied) can be passed here. locate() will pass relevant arguments to aggregate and method based on the names of additional arguments provided.

threading

a bool indicating whether to use the multi-threading to accelerate the criterion calculation. Turning this option on could improve the speed of criterion calculations when the emulator is built with a moderately large number of training data points and the Matérn-2.5 kernel.

aggregate

an R function that aggregates scores of the criterion specified by method across different output dimensions (if object is an instance of the dgp class) or across different emulators (if object is an instance of the bundle class). The function should be specified in the following basic form:

the first argument is a matrix representing scores. The rows of the matrix correspond to different design points in the candidate set. The number of columns of the matrix equals to:

the emulator output dimension if object is an instance of the dgp class; or

the number of emulators contained in object if object is an instance of the bundle class.

the output should be a vector that gives aggregations of scores at different design points.

Set to NULL to disable the aggregation. Defaults to NULL.

Value

If object is an instance of the gp class, a list is returned:

when x_cand = NULL, the list has one slot called location that contains

a single-row matrix that gives the position of determined design point, if batch_size = 1; or

a number of slots named position1,...,positionS where S = batch_size, each of which is a single-row matrix giving the position of a determined design point, if batch_size is greater than 1.

when x_cand is not NULL, the list has an additional slot called index that contains

the row index of the next design point in x_cand, if batch_size = 1; or

a number of slots named position1,...,positionS giving indices of next design points in x_cand, if batch_size is greater than 1.

If object is an instance of the dgp class, a list is returned:

when x_cand = NULL and aggregate = NULL, the list has one slot called location that contains

a matrix whose rows give the positions of determined design points with respect to different output dimensions, if batch_size = 1; or

a number of slots named position1,...,positionS where S = batch_size, each of which is a matrix giving the position of determined design points with respect to different output dimensions, if batch_size is greater than 1.

when x_cand is not NULL, the list has an additional slot called index that contains

a vector giving the row indices of determined design points in x_cand with respect to different output dimensions, if batch_size = 1; or

a number of vectors named position1,...,positionS, each of which gives indices of determined design points in x_cand with respect to different output dimensions, if batch_size is greater than 1.

when x_cand = NULL and aggregate is provided, the list has one slot called location that contains

a single-row matrix that gives the position of the determined design point, if batch_size = 1; or

a number of slots named position1,...,positionS where S = batch_size, each of which is a single-row matrix giving the position of the determined design point, if batch_size is greater than 1.

when x_cand is not NULL, the list has an additional slot called index that contains

the row index of the next design point in x_cand, if batch_size = 1; or

a number of slots named position1,...,positionS, each of which gives the index of a determined design point in x_cand, if batch_size is greater than 1.

If object is an instance of the bundle class, a list is returned:

when x_cand = NULL and aggregate = NULL, the list has one slot called location that contains

a matrix whose rows give the positions of determined design points with respect to different emulators in the bundle, if batch_size = 1;

a number of slots named position1,...,positionS where S = batch_size, each of which is a matrix giving the position of determined design points with respect to different emulators in the bundle, if batch_size is greater than 1.

when x_cand is not NULL, the list has an additional slot called index that contains

a vector giving the row indices of determined design points in x_cand with respect to different emulators in the bundle, if batch_size = 1; or

a number of vectors named position1,...,positionS, each of which gives indices of determined design points in x_cand with respect to different emulators in the bundle, if batch_size is greater than 1.

when x_cand = NULL and aggregate is provided, the list has one slot called location that contains

a single-row matrix that gives the position of the determined design point, if batch_size = 1; or

a number of slots named position1,...,positionS where S = batch_size, each of which is a single-row matrix giving the position of the determined design point, if batch_size is greater than 1.

when x_cand is not NULL, the list has an additional slot called index that contains

the row index of the next design point in x_cand, if batch_size = 1; or

a number of slots named position1,...,positionS, each of which gives the index of a determined design point in x_cand, if batch_size is greater than 1.

Details

See further examples and tutorials at https://mingdeyu.github.io/dgpsi-R/.

Note

The function is only applicable to GP emulators, DGP emulators without likelihood layers, or bundles of (D)GP emulators created by pack().

Any R vector detected in x_cand will be treated as a column vector and automatically converted into a single-column R matrix.

References

MacKay, D. J. (1992). Information-based objective functions for active data selection. Neural Computation, 4(4), 590-604.

Beck, J., & Guillas, S. (2016). Sequential design with mutual information for computer experiments (MICE): emulation of a tsunami model. SIAM/ASA Journal on Uncertainty Quantification, 4(1), 739-766.

Examples

if (FALSE) {

# load packages and the Python env
library(lhs)
library(dgpsi)
init_py()

# construct a 1D non-stationary function
f <- function(x) {
 sin(30*((2*x-1)/2-0.4)^5)*cos(20*((2*x-1)/2-0.4))
}

# generate the initial design
X <- maximinLHS(10,1)
Y <- f(X)

# training a 2-layered DGP emulator with the global connection off
m <- dgp(X, Y, connect = F)

# specify the range of the input dimension
lim <- c(0,1)

# locate the next design point
next_point <- locate(m, limits = lim)
X_new <- next_point$location

# obtain the corresponding output at the located design point
Y_new <- f(X_new)

# combine the new input-output pair to the existing data
X <- rbind(X, X_new)
Y <- rbind(Y, Y_new)

# update the DGP emulator with the new input and output data and refit with 500 training iterations
m <- update(m, X, Y, refit = TRUE, N = 500)

# plot the LOO validation
plot(m)
}