Summary of a constructed GP, DGP, or linked (D)GP emulator

This function provides a summary of key information for a GP, DGP, or linked (D)GP emulator by generating either a table or an interactive plot of the emulator’s structure.

Usage

# S3 method for class 'gp'
summary(object, type = "plot", ...)

# S3 method for class 'dgp'
summary(object, type = "plot", ...)

# S3 method for class 'lgp'
summary(object, type = "plot", group_size = 1, ...)

Arguments

object

can be one of the following:

the S3 class gp.

the S3 class dgp.

the S3 class lgp.

type

a character string, either "table" or "plot", indicating the format of the output. If set to "table", the function returns a summary in table. If set to "plot", the function returns an interactive visualization. Defaults to "plot". If the object was created with lgp() where struc is not a data frame, type will automatically default to "table".

...

N/A.

group_size

an integer specifying the number of consecutive layers to be grouped together in the interactive visualization of linked emulators when type = "plot". This argument is only applicable if object is an instance of the lgp class. Defaults to 1.

Value

Either a table or an interactive visualization of the emulator, returned as a visNetwork object. The visualization is compatible with R Markdown documents and the RStudio Viewer. The resulting visNetwork object can be saved as an HTML file using visNetwork::visSave().

Details

See further examples and tutorials at https://mingdeyu.github.io/dgpsi-R/.

Examples

if (FALSE) { # \dontrun{

# See gp(), dgp(), or lgp() for an example.
} # }